Revert "chores: flashcards list"
