+++ src/App.css
@@ -13,14 +13,3 @@ body {
   
 }
 
-.cardList_container{
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  overflow-y: auto;
-}
-
-.cardList_container > * {
-  margin-top: 12px;
-}
