+++ src/components/AppHeader.tsx
@@ -4,11 +4,9 @@ import logotype from "../assets/header_logo.svg";
 import button from "../assets/header_button.svg";
 
 export interface Header {
-  currentCardsNumber: number;
+  CurrentCardsNumber: number;
   changeMode: (event: React.MouseEvent<HTMLButtonElement>) => void;
 }
-
-
 
 export function AppHeader(props: Header){
 
@@ -16,7 +14,7 @@ export function AppHeader(props: Header){
     <div>
         <img src={logotype} alt={"logo"} />
         <h1 className="header">Cards:</h1>
-        <span className="cardsNumber">{props.currentCardsNumber}</span>
+        <span className="cardsNumber">{props.CurrentCardsNumber}</span>
     </div >
      <button className="addButton" onClick={props.changeMode}>
       <img src={button} alt={"add card button"} />
