+++ src/components/AppLayout.module.css
@@ -3,6 +3,4 @@
   height: 755px;
   margin: 1% auto;
   background: #f1f2f3;
-  display: flex;
-  flex-direction: column;
 }
