+++ src/components/Card.tsx
@@ -1,59 +1,86 @@
-import React, { useRef, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import styles from "./Card.module.css";
-import editIcon from "../assets/editIcon.svg";
-
+import deleteImage from "./assets/deleteIcon.svg";
+import { TextareaInput } from "./TextareaInput";
 
 interface FishkappCard {
   question: string;
   answer: string;
 }
-
 interface CardI {
-    id: number;
-    question: string;
-    answer: string;
+  editMode: boolean;
+  setEditMode: (value: boolean) => void;
 }
 
 export const Card = (props: CardI) => {
-    const [nextPage, setNextPage] = useState(false);
-    const textareaRef = useRef<HTMLTextAreaElement>(null);
-    const [height, setHeight] = useState<number>(60);
-  
-    const [fishkappObject, setfishkappObject] = useState<FishkappCard>({
-      question: props.question ? props.question : "",
-      answer: props.answer ? props.answer : "",
-    });
+  const [nextPage, setNextPage] = useState(false);
+  const textareaRef = useRef<HTMLTextAreaElement>(null);
 
+  const [fishkappObject, setfishkappObject] = useState<FishkappCard>({
+    question: "",
+    answer: "",
+  });
 
-  const changePageClick = () => {
-    setNextPage(!nextPage);
+  const handleInputChange = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
+    const { value } = event.target;
+    nextPage
+      ? setfishkappObject({ ...fishkappObject, answer: value })
+      : setfishkappObject({ ...fishkappObject, question: value });
+    event.target.style.height = "0px";
+    event.target.style.height = event.target.scrollHeight + "px";
+  };
+
+  const nextPageClick = () => {
+    setNextPage(true);
     textareaRef.current?.focus();
+  };
+  const backPageClick = () => {
+    setNextPage(false);
+    textareaRef.current?.focus();
+  };
+  const cancelPageClick = () => {
+    props.setEditMode(!props.editMode);
   };
 
   return (
     <div className={styles.container}>
       <div className={styles.corner_wrapper}>
-                <button className={styles.corner_button}>
-                    <img src={editIcon} alt="edit" />
-                </button>
-      </div>
-
-      <div className={styles.text_wrapper} onClick={changePageClick}>
         {nextPage ? (
-            <textarea
-            readOnly
-            className={styles.output}
-            value={fishkappObject.answer}
-            style={{ height: height + "px" }}
-            />
-        ):
-        (
-            <textarea
-            readOnly
-            className={styles.output}
-            value={fishkappObject.question}
-            style={{ height: height + "px" }}
-            />
+          <button className={styles.corner_button}>
+            <img src={deleteImage} alt="delete" />
+          </button>
+        ) : (
+          <></>
+        )}
+      </div>
+      <div className={styles.text_wrapper}>
+        {nextPage && (
+          <p className={styles.question_text}>{fishkappObject.question}</p>
+        )}
+
+        <TextareaInput
+          fishkappObject={fishkappObject}
+          nextPage={nextPage}
+          handleInputChange={handleInputChange}
+          ref={textareaRef}
+        />
+      </div>
+      <div className={styles.action_wrapper}>
+        {nextPage ? (
+          <button onClick={backPageClick} className={styles.left_button}>
+            Back
+          </button>
+        ) : (
+          <button onClick={cancelPageClick} className={styles.left_button}>
+            Cancel
+          </button>
+        )}
+        {nextPage ? (
+          <button className={styles.right_button}>Save</button>
+        ) : (
+          <button onClick={nextPageClick} className={styles.right_button}>
+            Next
+          </button>
         )}
       </div>
     </div>
