--- src/components/NewCard.module.css
@@ -1,74 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    background-color: white;
-    width: 343px;
-    min-height: 188px;
-    position: relative;
-    border-radius: 15px;
-    border: 0.5px #ebebec solid;
-    justify-content: space-between;
-    align-items: center;
-  }
-  .corner_wrapper {
-    display: flex;
-    flex-direction: row;
-    width: 343px;
-    height: 40px;
-    justify-content: flex-end;
-  }
-  .corner_button {
-    position: relative;
-    background: transparent;
-    border: none;
-    width: 40px;
-    height: 40px;
-    cursor: pointer;
-  }
-  .text_wrapper {
-    position: relative;
-  }
-  .question_text {
-    position: relative;
-    bottom: 0;
-    color: #161616;
-    font-size: 12px;
-    font-family: Lexend;
-    max-width: 300px;
-    font-weight: 400;
-    word-wrap: break-word;
-    margin: 0 15px 0;
-  }
-  .action_wrapper {
-    display: flex;
-    flex-direction: row;
-    position: absolute;
-    align-self: end;
-    gap: 10px;
-    bottom: 14px;
-    right: 14px;
-  }
-  .left_button {
-    width: 110px;
-    height: 38px;
-    padding: 10px, 12px, 10px, 12px;
-    border: none;
-    border-radius: 22px;
-    background-color: #ffffff;
-    line-height: 18px;
-    word-wrap: break-word;
-    font-weight: 500;
-    cursor: pointer;
-  }
-  .right_button {
-    width: 110px;
-    height: 38px;
-    border: none;
-    border-radius: 22px;
-    background-color: #161616;
-    color: white;
-    line-height: 18px;
-    word-wrap: break-word;
-    font-weight: 500;
-    cursor: pointer;
-  }
